* UPD: Refactoring InGameTime System
* MOVE: Moved InGameTime System into xrEngine
