+++ code/engine.vc2008/FrayBuildConfig.hpp
@@ -16,10 +16,10 @@
 //#define TEST_TTAPI			// Perform a test of ttapi integrity
 
 /* xrCore */
-#define	LOG_TIME_PRECISE		// (by alpet & RvP)
+#define	LOG_TIME_PRECISE		// Optimization log writte (by alpet & RvP)
 
 /* Scripts */
-//#define SPAWN_ANTIFREEZE		//* spread spawn of game objects thoughout multiple frames to prevent lags (by alpet) */
+//#define SPAWN_ANTIFREEZE		// spread spawn of game objects thoughout multiple frames to prevent lags
 //#define LUACP_API				// attaching luaicp.dll
 
 /* xrRender */
@@ -27,18 +27,18 @@
 
 /* xrGame */
 #define WPN_AUTORELOAD			// Enable autoreload
-#define DYNAMIC_SUN_MOVEMENT 	// Use dynamic sun movement
+//#define DYNAMIC_SUN_MOVEMENT 	// Use dynamic sun movement
 //#define NEW_SLOTS				// Add knife slot and binocular slot in inventory
 //#define ASPAWN_CROW			// Spawn crows over actor
 //#define NEW_ANIMS_WPN			// Enable anm_..._crouch, anm_reload_empty
 #define DEAD_BODY_WEAPON		// Enable weapons in dead stalkers
-//#define POLTERGEIST_BUILD_DIE	// Old style polter dead
+#define POLTERGEIST_BUILD_DIE	// Old style polter dead. Need fixes for scripts and configs.
 
-/*		Inventory/HUD states value */
+/* IUI states value */
 //#define ARROW_VALUE_INV		// Adds arrows to inventory. Clear Sky style. For use unlock in actor_menu/_16.xml tags <arrow> and <arrow_shadow>
 #define PROGRESS_VALUE_INV		// Adds progress bars value in the inventory. COP style. Used for default values.
 //#define NUM_PARAMS_INV		// Adds progress bars in the inventory numeric value. Max value 100%. (%) - color="no_color"
 //#define SHAPE_VALUE_INV		// Adds shape progress to inventory. 
 //#define MINMAP_BAR_HUD		// Adds progress bars to luminosity and noise value
 //#define HUD_DOSIMETER			// Adds dosimeter on hud. From Clear Sky.
-#define HUD_SHAPE_VALUE			// Adds shape progress to HUD. Health and Stamina value.
+//#define HUD_SHAPE_VALUE		// Adds shape progress to HUD. Health and Stamina value.
