+++ code/engine.vc2008/xrGame/Level.cpp
@@ -21,7 +21,7 @@
 #include "script_engine_space.h"
 #include "infoportion.h"
 #include "patrol_path_storage.h"
-#include "date_time.h"
+#include "../xrEngine/date_time.h"
 #include "space_restriction_manager.h"
 #include "seniority_hierarchy_holder.h"
 #include "space_restrictor.h"
