+++ code/engine.vc2008/xrGame/alife_time_manager.cpp
@@ -8,7 +8,7 @@
 
 #include "stdafx.h"
 #include "alife_time_manager.h"
-#include "date_time.h"
+#include "../xrEngine/date_time.h"
 
 CALifeTimeManager::CALifeTimeManager	(LPCSTR section)
 {
