+++ code/engine.vc2008/xrGame/autosave_manager.cpp
@@ -8,7 +8,7 @@
 
 #include "stdafx.h"
 #include "autosave_manager.h"
-#include "date_time.h"
+#include "../xrEngine/date_time.h"
 #include "ai_space.h"
 #include "level.h"
 #include "xrMessages.h"
