+++ code/engine.vc2008/xrGame/console_commands.cpp
@@ -26,7 +26,7 @@
 #include "autosave_manager.h"
 #include "ai_space.h"
 #include "ai/monsters/BaseMonster/base_monster.h"
-#include "date_time.h"
+#include "../xrEngine/date_time.h"
 #include "UIGameSP.h"
 #include "ui/UIActorMenu.h"
 #include "ui/UIStatic.h"
