+++ code/engine.vc2008/xrGame/game_news.cpp
@@ -12,7 +12,7 @@
 #include "alife_object_registry.h"
 #include "game_graph.h"
 
-#include "date_time.h"
+#include "../xrEngine/date_time.h"
 #include "xrServer_Objects_ALife_Monsters.h"
 
 #include "specific_character.h"
