+++ code/engine.vc2008/xrGame/script_game_object_inventory_owner.cpp
@@ -12,7 +12,7 @@
 #include "gametask.h"
 #include "actor.h"
 #include "level.h"
-#include "date_time.h"
+#include "../xrEngine/date_time.h"
 #include "uigamesp.h"
 #include "restricted_object.h"
 #include "script_engine.h"
