+++ code/engine.vc2008/xrGame/ui/UIInventoryUtilities.cpp
@@ -1,4 +1,3 @@
-
 #include "stdafx.h"
 #include "UIInventoryUtilities.h"
 #include "../WeaponAmmo.h"
@@ -6,7 +5,7 @@
 #include "UIStatic.h"
 #include "../eatable_item.h"
 #include "../Level.h"
-#include "../date_time.h"
+#include "../xrEngine/date_time.h"
 #include "../string_table.h"
 #include "../Inventory.h"
 #include "../InventoryOwner.h"
