+++ code/engine.vc2008/xrGame/ui/UIMainIngameWnd.cpp
@@ -4,7 +4,6 @@
 #include "UIMessagesWindow.h"
 #include "../UIZoneMap.h"
 
-
 #include <dinput.h>
 #include "../actor.h"
 #include "../ActorCondition.h"
@@ -25,7 +24,7 @@
 #include "../level.h"
 #include "../seniority_hierarchy_holder.h"
 
-#include "../date_time.h"
+#include "../xrEngine/date_time.h"
 #include "../xrServerEntities/xrServer_Objects_ALife_Monsters.h"
 #include "../../xrEngine/LightAnimLibrary.h"
