+++ code/engine.vc2008/xrGame/xr_time.cpp
@@ -2,7 +2,7 @@
 #include "xr_time.h"
 #include "ui/UIInventoryUtilities.h"
 #include "level.h"
-#include "date_time.h"
+#include "../xrEngine/date_time.h"
 #include "ai_space.h"
 #include "alife_simulator.h"
 #include "alife_time_manager.h"
